requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
h11==0.11.0
idna==2.10
mariadb==1.0.4
orjson==3.4.3
pydantic==1.7.2
PyJWT==1.7.1
redis==3.5.3
requests==2.24.0
starlette==0.13.6
urllib3==1.25.11
uvicorn==0.12.2
